+++ src/main/java/fr/inra/oresing/workflow/cascade/StagingFinalizeSql.java
@@ -444,7 +444,7 @@ public final class StagingFinalizeSql {
                 totalAffected += affected;
                 if (affected > 0) {
                     try {
-                        onBatchUpserted.accept((long) affected);
+                        onBatchUpserted.accept(affected);
                     } catch (RuntimeException ignored) {
                         /* best effort : un consommateur fautif ne doit pas
                            casser le UPSERT en cours */
